Line of Sight Visualizer: Init
This is a partial rewrite of https://rop.nl/freifunk/line-of-sight.php. Webscraping got replaced by API-Calls. The main functionality is working, however more detailed information cant be shown as long as there are not available as attributes in the wiki. I consider this optional as you can still get all the relevant information in the wikiarticle, so its not strictly needed in Google Earth.